Fix visibility issue with Kotlin (#65)
* Extract two public static inner classes

  - These classes cause an issue with Kotlin
    (https://youtrack.jetbrains.com/issue/KT-11398)
